pyproject.toml
[tool.poetry]
name = "task_manager"
version = "0.1.0"
description = ""
authors = ["Kauan Azevedo <kauanazevedo2010@gmail.com>"]

[tool.poetry.scripts]
django = "django:main"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.13"
Django = "^4.2.7"
python-decouple = "^3.8"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
